+++ CHANGELOG.md
@@ -33,6 +33,7 @@
 * 【json  】      修复JSONBeanParser在遇到List时没有被正确递归问题（issue#I7M2GZ@Gitee）
 * 【core  】      修复VersionComparator对1.0.3及1.0.2a比较有误的问题（pr#1043@Gitee）
 * 【core  】      修复IOS系统下，chrome 浏览器的解析规则有误（pr#1044@Gitee）
+* 【extra 】      修复多线程下Sftp中Channel关闭的问题（issue#I7OHIB@Gitee）
 
 -------------------------------------------------------------------------------------------------------------
 # 5.8.20(2023-06-16)